requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
diffusers>=0.25.0
kornia>=0.7.0
vector-quantize-pytorch>=1.12.0
einops>=0.7.0
accelerate>=0.25.0
pyyaml
tqdm
mamba-ssm>=1.1.1